feat: (Firebase) - rename SyncMessagesWorker.kt to SyncMailboxesWorker.kt
